+++ src/handlers/calculateHash.js
