+++ src/handlers/rm.js
