+++ src/helpers/setWorkingDirectory.js
